MapReduce: enforce the modulo 3 per card instead of per page
Some cards may get lost, defeating our verification.

To make it work, I had to go from the rule XX % 3 == 0 to XX % 3 == 1,
and remove one pattern.

I tried to play alone, and 4 patterns seem better to me anyway.

I also changed the generation algorithm so that the patterns counts
are not always the same on each card. The way I'm doing for skewed is
.. somewhat cheated but I dunno how to do it properly.